pyproject.toml
[tool.poetry]
name = "taverntales"
version = "0.1.0"
description = ""
authors = ["duysqubix <duys@qubixds.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
openai = "^1.59.2"
elevenlabs = "^1.50.3"
anvil-uplink = "^0.5.2"
jupyter = "^1.1.1"
sqlalchemy = "^2.0.36"
django = "^5.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
